Remove platform definitions.
Platform definitions will now be kept in separate repositories.

Auto-detection will be used if the apollo_fpga module is present, and
the detected Apollo device is supported by the installed 'cynthion' or
'luna_boards' modules.

In all other cases, the LUNA_PLATFORM environment variable must be set
for get_appropriate_platform() to return a platform.